Create SQL Code For Database
This is where I am going to be dropping off all of the code I work on. Right now is just a bunch of tables made to receive and hold data. I have been planning out flow of data and need to ask around about what functions we need as a lot of administrative processes deal with the data.