Update SQL Code For Database
This was the code to show I implemented a table for majors and then individual tables for each majors requirements. I also had to go through and populate the requirements tables. For this I needed some help from AI as there was a lot of data. I rearranged the data into CSV. Then wrote some code for the first one. Then used AI to just modulate the inserts over and over again.